MovieVerse-Mobile/platforms/android/app/src/main/assets/www/MovieVerse-Backend/databases/requirements.txt
requests==2.32.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.7.2
redis==5.0.4
kafka-python==2.0.2
pika==1.3.2
opensearch-py==2.4.2
